+++ packages/fastboard/src/index.ts
@@ -17,6 +17,9 @@ export type { Theme, Language, GenericIcon };
 
 export type MountProps = Omit<FastboardProps, "app">;
 
+/**
+ * @deprecated Use `createUI` instead.
+ */
 export function mount(app: FastboardApp, div: HTMLDivElement, options?: MountProps) {
   const fastboard = new Fastboard({ target: div, props: { app, ...options } });
 
@@ -32,6 +35,9 @@ export function mount(app: FastboardApp, div: HTMLDivElement, options?: MountPro
 
 export type ReplayProps = Omit<ReplayFastboardProps, "player">;
 
+/**
+ * @deprecated Use `createReplayUI` instead.
+ */
 export function replay(player: FastboardPlayer, div: HTMLDivElement, options?: ReplayProps) {
   const replayFastboard = new ReplayFastboard({ target: div, props: { player, ...options } });